Initial Commit
Player movement is basic, flawed, but it works.  The x-coordinate
movement should probably be split up into KeyPressed and KeyReleased.
Perhaps the same should be done for jumping.

The image filepaths are hardcoded for my computer.  That should probably
be fixed very soon.

I'll continue working on this later in the week before meeting on
Saturday.